Now we identify which container a sign is attached to
